+++ rgatCore/BridgeConnection.cs
@@ -994,8 +994,6 @@ namespace rgat
 
                 foreach (var item in items)
                 {
-                    if (rgatState.NetworkBridge.GUIMode)
-                        Console.WriteLine($"Sending raw {ASCIIEncoding.ASCII.GetString(item.data)}");
                     if (!RawSendData(item))
                     {
                         Teardown("Send outgoing loop failed");